+++ src/components/Header/Header.tsx
@@ -4,7 +4,6 @@ import * as React from 'react'
 import { childrenExist, customPropTypes, UIComponent } from '../../lib'
 import HeaderDescription from './HeaderDescription'
 import headerRules from './headerRules'
-import headerVariables from './headerVariables'
 
 /**
  * A header provides a short summary of content
@@ -44,8 +43,6 @@ class Header extends UIComponent<any, any> {
   static handledProps = ['as', 'children', 'className', 'content', 'description', 'textAlign']
 
   static rules = headerRules
-
-  static variables = headerVariables
 
   static Description = HeaderDescription