+++ src/components/Menu/Menu.tsx
@@ -7,22 +7,7 @@ import MenuItem from './MenuItem'
 import menuRules from './menuRules'
 import menuVariables from './menuVariables'
 
-export type MenuType = 'primary' | 'secondary'
-export type MenuShape = 'pills' | 'pointing' | 'underlined'
-
-export interface IMenuProps {
-  as?: string
-  activeIndex?: number | string
-  children?: React.ReactNode
-  className?: string
-  defaultActiveIndex?: number | string
-  items?: any
-  shape?: MenuShape
-  type?: MenuType
-  vertical?: boolean
-}
-
-class Menu extends AutoControlledComponent<IMenuProps, any> {
+class Menu extends AutoControlledComponent<any, any> {
   static displayName = 'Menu'
 
   static className = 'ui-menu'