+++ src/components/Menu/MenuItem.tsx
@@ -4,26 +4,10 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
 import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
-
-import { MenuType, MenuShape } from './Menu'
-
 import menuItemRules from './menuItemRules'
 import menuVariables from './menuVariables'
 
-export interface IMenuItemProps {
-  active?: boolean
-  as?: string
-  children?: React.ReactNode
-  className?: string
-  content?: React.ReactNode
-  index?: number
-  onClick?: (any, IMenuItemProps) => void
-  shape?: MenuShape
-  type?: MenuType
-  vertical?: boolean
-}
-
-class MenuItem extends UIComponent<IMenuItemProps, any> {
+class MenuItem extends UIComponent<any, any> {
   static displayName = 'MenuItem'
 
   static className = 'ui-menu__item'